pyproject.toml
[tool]
[tool.poetry]
name = "blackboard-pagi"
version = "0.1.0"
homepage = "https://github.com/blackhc/blackboard-pagi"
description = "Proto-AGI using a Blackboard System (for the LLM Hackathon by Ben's Bites)."
authors = ["Andreas Kirsch <blackhc@gmail.com>"]
readme = "README.md"
license =  "GPL-3.0-only"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.11',
]
packages = [
    { include = "blackboard_pagi" },
    { include = "tests", format = "sdist" },
    # Include all files under blackboard_pagi/blackboard/template
    { include = "blackboard_pagi/blackboard/template/**/*" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
click = "8.0.1"

black  = { version = "^22.12.0", optional = true}
isort  = { version = "^5.8.0", optional = true}
flake8  = "^6.0.0"
flake8-docstrings = { version = "^1.6.0", optional = true }
mypy = "^1.1.1"
pytest  = { version = "^6.2.4", optional = true}
pytest-cov  = { version = "^2.12.0", optional = true}
tox  = { version = "^3.20.1", optional = true}
mkdocs  = { version = "^1.4.2", optional = true}
mkdocs-include-markdown-plugin  = { version = "^1.0.0", optional = true}
mkdocs-material  = { version = "^6.1.7", optional = true}
mkdocstrings  = {version = "^0.19.0", optional = true, extras=["python"]}
mkdocs-material-extensions  = { version = "^1.1.1", optional = true}
twine  = { version = "^3.3.0", optional = true}
mkdocs-autorefs = { version = "^0.4.1", optional = true}
pre-commit = {version = "^2.12.0", optional = true}
toml = {version = "^0.10.2", optional = true}
bump2version = {version = "^1.0.1", optional = true}
jinja2 = {version = "<3.1.0", optional = false}
blackhc-project = "^2.6.0"
langchain = {version = "^0.0.126"}
ipykernel = {version = "^6.19.2", optional = true}
gitpython = "^3.1.29"
pip = "^22.3.1"
openai = "^0.27.0"
manifest-ml = "^0.0.1"
python-frontmatter = "^1.0.0"
matplotlib = "^3.6.2"
pillow = "^9.3.0"
parse = {version = "^1.19.0"}
docstring-parser = "^0.15"
dacite = "^1.6.0"
jax = {version = "^0.4.5", extras = ["cpu"]}
llm-strategy = {version = ">=0.0.3"}
markdown = "<3.4"
pydantic = "^1.10.7"

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "flake8",
    "flake8-docstrings",
    "pytest-cov"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version", "ipykernel"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs"
    ]

[tool.poetry.scripts]
blackboard-pagi = 'blackboard_pagi.cli:main'

[tool.poetry.group.dev.dependencies]
parse = {version = "^1.19.0", optional = true}

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
python_files = "test_*.py"
